[added] enahance med packs to bluefrog [added] /healenhace [added] /grantskill <skill> see skill table for skillbox names [fixed] transferitem, spatialchat priority
git-svn-id: http://svn2.assembla.com/svn/swgemu/core3/trunk/MMOCoreORB@2304 c3d1530f-68f5-4bd0-87dc-8ef779617e40

+++ src/server/zone/templates/tangible/RangedStimPackTemplate.h
@@ -34,7 +34,7 @@ public:
 		range = templateData->getFloatField("range");
 		rangeMod = templateData->getFloatField("rangeMod");
 		area = templateData->getFloatField("area");
-		effectiveness = templateData->getIntField("effectiveness");
+		effectiveness = templateData->getFloatField("effectiveness");
     }
 
 	inline float getEffectiveness() {
